% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appe.glm.R
\name{appe.glm}
\alias{appe.glm}
\title{C-statistics adjusted for predictor distributions
Calculates adjusted \eqn{C} statistics by predictor distributions for a generalized linear model with binary outcome.}
\usage{
appe.glm(mdl, dat.train, dat.test, method = "uLSIF", sigma = NULL,
  lambda = NULL, kernel_num = NULL, fold = 5, stabilize = TRUE,
  qstb = 0.025, reps = 2000, conf.level = 0.95)
}
\arguments{
\item{mdl}{A glm object describing a prediction model to be evaluated.}

\item{dat.train}{A dataframe used to construct a prediction model (specified in mdl), corresponding to a training data. Need to include outcome and all predictors.}

\item{dat.test}{A dataframe corresponding to a validation (testing) data. Need to include outcome and all predictors.}

\item{method}{uLSIF or KLIEP. Same as the argument in densratio function from densratio package.}

\item{sigma}{A positive numeric vector corresponding to candidate values of a bandwidth for Gaussian kernel. Same as the argument in densratio function from densratio package.}

\item{lambda}{A positive numeric vector corresponding to candidate values of a regularization parameter. Same as the argument in densratio function from densratio package.}

\item{kernel_num}{A positive integer corresponding to number of kernels. Same as the argument in densratio function from densratio package.}

\item{fold}{A positive integer corresponding to a number of the folds of cross-validation in the KLIEP method. Same as the argument in densratio function from densratio package.}

\item{stabilize}{A logical value as to whether tail weight stabilization is performed or not. If TRUE, both tails of the estimated density ratio distribution are replaced by the constant value which is specified at qstb option.}

\item{qstb}{A positive numerical value less than 1 to control the degree of weight stabilization. Default value is 0.025, indicating estimated density ratio values less than the 2.5 percentile and more than the 97.5 percentile are set to 2.5 percentile and 97.5 percentile, respectively.}

\item{reps}{A positive integer to specify bootstrap repetitions. If 0, bootstrap calculations are not performed.}

\item{conf.level}{A numerical value indicating a confidence level of interval.}
}
\value{
matrix
}
\description{
C-statistics adjusted for predictor distributions
Calculates adjusted \eqn{C} statistics by predictor distributions for a generalized linear model with binary outcome.
}
