% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalest.bin.R
\name{cvalest.bin}
\alias{cvalest.bin}
\title{C-statistics adjusted for predictor distributions
Estimation of C-statistics
Calculates C-statistics. Individual case  weight can be incorporated.}
\usage{
cvalest.bin(Y, scr, wgt = NULL)
}
\arguments{
\item{Y}{A numerical vector of inary outcome, either 0 or 1.}

\item{scr}{A numerical vector of continuous variable.}

\item{wgt}{A numerical vector corresponding to individuatl weight.}
}
\value{
c-statistics
}
\description{
C-statistics adjusted for predictor distributions
Estimation of C-statistics
Calculates C-statistics. Individual case  weight can be incorporated.
}
